requirements.txt
Django==4.1.5
Pillow==9.4.0
psycopg2==2.9.5
djangorestframework==3.14.0